nvim: Refactor the config into modules
This commit completely restructures the architecture of my neovim
config. It separates the files between a "core" folder and a "modules"
folder.

The "core" folder is always expected to be loaded.
The "modules" folder contains different modules, which I split according
to what their target usage was (ui, completion, lsp, etc...)

Note that a module _can_ depend on another one. _e.g._ The colour scheme
expects the tree-sitter module to be loaded. Loading the "ui" module
will load the treesitter module automatically.